+++ couchdb/couchdb_test.go
@@ -12,11 +12,12 @@ package couchdb_test
 //--------------------
 
 import (
-	"fmt"
+	"strings"
 	"testing"
 
 	"github.com/tideland/golib/audit"
 	"github.com/tideland/golib/etc"
+	"github.com/tideland/golib/identifier"
 
 	"github.com/tideland/gocouch/couchdb"
 )
@@ -30,6 +31,7 @@ const (
 	LocalhostCfg   = "{etc {hostname localhost}{port 5984}}"
 	TestingDBCfg   = "{etc {hostname localhost}{port 5984}{database tgocouch-testing}}"
 	TemporaryDBCfg = "{etc {hostname localhost}{port 5984}{database tgocouch-testing-temporary}}"
+	TemplateDBcfg  = "{etc {hostname localhost}{port 5984}{database tgocouch-testing-<<DATABASE>>}}"
 )
 
 //--------------------
@@ -67,27 +69,33 @@ func TestCreateDeleteDatabase(t *testing.T) {
 	cfg, err := etc.ReadString(TemporaryDBCfg)
 	assert.Nil(err)
 
+	// Open and check existance.
 	cdb, err := couchdb.Open(cfg)
 	assert.Nil(err)
-	ids, err := cdb.AllDatabases()
+	has, err := cdb.HasDatabase()
 	assert.Nil(err)
-	dbNo := len(ids)
+	assert.False(has)
+
+	// Create and check existance,
 	resp := cdb.CreateDatabase()
 	assert.True(resp.IsOK())
-	ids, err = cdb.AllDatabases()
+	has, err = cdb.HasDatabase()
 	assert.Nil(err)
-	assert.Equal(len(ids), dbNo+1)
+	assert.True(has)
+
+	// Delete and check existance.
 	resp = cdb.DeleteDatabase()
 	assert.True(resp.IsOK())
-	ids, err = cdb.AllDatabases()
+	has, err = cdb.HasDatabase()
 	assert.Nil(err)
-	assert.Equal(len(ids), dbNo)
+	assert.False(has)
 }
 
 // TestCreateDesignDocument tests creating new design documents.
 func TestCreateDesignDocument(t *testing.T) {
 	assert := audit.NewTestingAssertion(t, true)
-	cdb := prepareDatabase(assert)
+	cdb, cleanup := prepareFilledDatabase("create-design", assert)
+	defer cleanup()
 
 	// Create design document and check if it has been created.
 	allDesignA, err := cdb.AllDesignDocuments()
@@ -95,8 +103,8 @@ func TestCreateDesignDocument(t *testing.T) {
 
 	ddocA := &couchdb.DesignDocument{
 		ID: "_design/testing",
-		Views: couchdb.Views{
-			"foo": couchdb.View{
+		Views: couchdb.DesignViews{
+			"index-a": couchdb.DesignView{
 				Map: "function(doc){ if (doc._id.indexOf('a') !== -1) { emit(doc._id, doc._rev);  } }",
 			},
 		},
@@ -112,13 +120,14 @@ func TestCreateDesignDocument(t *testing.T) {
 // TestViewDocuments tests calling a view.
 func TestViewDocuments(t *testing.T) {
 	assert := audit.NewTestingAssertion(t, true)
-	cdb := prepareDatabase(assert)
+	cdb, cleanup := prepareFilledDatabase("view-documents", assert)
+	defer cleanup()
 
 	// Create design document.
 	ddocA := &couchdb.DesignDocument{
 		ID: "_design/testing",
-		Views: couchdb.Views{
-			"aid": couchdb.View{
+		Views: couchdb.DesignViews{
+			"index-a": couchdb.DesignView{
 				Map: "function(doc){ if (doc._id.indexOf('a') !== -1) { emit(doc._id, doc._rev);  } }",
 			},
 		},
@@ -126,37 +135,57 @@ func TestViewDocuments(t *testing.T) {
 	resp := cdb.CreateDesignDocument(ddocA)
 	assert.True(resp.IsOK())
 
-	// Now add some test data.
-	gen := audit.NewGenerator(audit.SimpleRand())
-	for i := 0; i < 100; i++ {
-		doc := MyDocument{
-			DocumentID: fmt.Sprintf("%s-%d", gen.Word(), i),
-			FieldA:     gen.Sentence(),
-			FieldB:     gen.Int(1, 999),
-		}
-		resp = cdb.CreateDocument(doc)
-		assert.True(resp.IsOK())
-	}
-
-	// Last but not least call the view.
-	resp = cdb.ViewDocuments("testing", "aid")
+	// Call the view for the first time.
+	resp = cdb.ViewDocuments("testing", "index-a")
 	assert.True(resp.IsOK())
 	vr := couchdb.ViewResult{}
 	err := resp.ResultValue(&vr)
 	assert.Nil(err)
-	assert.True(vr.TotalRows > 0)
-	assert.Equal(vr.Offset, 0)
+	trOld := vr.TotalRows
+	assert.True(trOld > 0)
+
+	// Add a matching document and view again.
+	docA := MyDocument{
+		DocumentID: "black-jack-4711",
+		Name:       "Jack Black",
+	}
+	resp = cdb.CreateDocument(docA)
+	assert.True(resp.IsOK())
+	resp = cdb.ViewDocuments("testing", "index-a")
+	assert.True(resp.IsOK())
+	vr = couchdb.ViewResult{}
+	err = resp.ResultValue(&vr)
+	assert.Nil(err)
+	trNew := vr.TotalRows
+	assert.Equal(trNew, trOld+1)
+
+	// Add a non-matching document and view again.
+	docB := MyDocument{
+		DocumentID: "doe-john-999",
+		Name:       "John Doe",
+	}
+	resp = cdb.CreateDocument(docB)
+	assert.True(resp.IsOK())
+	resp = cdb.ViewDocuments("testing", "index-a")
+	assert.True(resp.IsOK())
+	vr = couchdb.ViewResult{}
+	err = resp.ResultValue(&vr)
+	assert.Nil(err)
+	trFinal := vr.TotalRows
+	assert.Equal(trFinal, trNew)
+
 }
 
 // TestCreateDocument tests creating new documents.
 func TestCreateDocument(t *testing.T) {
 	assert := audit.NewTestingAssertion(t, true)
-	cdb := prepareDatabase(assert)
+	cdb, cleanup := prepareDatabase("create-document", assert)
+	defer cleanup()
 
 	// Create document without ID.
 	docA := MyDocument{
-		FieldA: "foo",
-		FieldB: 4711,
+		Name: "foo",
+		Age:  50,
 	}
 	resp := cdb.CreateDocument(docA)
 	assert.True(resp.IsOK())
@@ -166,8 +195,9 @@ func TestCreateDocument(t *testing.T) {
 	// Create document with ID.
 	docB := MyDocument{
 		DocumentID: "bar-12345",
-		FieldA:     "bar",
-		FieldB:     12345,
+		Name:       "bar",
+		Age:        25,
+		Active:     true,
 	}
 	resp = cdb.CreateDocument(docB)
 	assert.True(resp.IsOK())
@@ -178,13 +208,14 @@ func TestCreateDocument(t *testing.T) {
 // TestReadDocument tests reading a document.
 func TestReadDocument(t *testing.T) {
 	assert := audit.NewTestingAssertion(t, true)
-	cdb := prepareDatabase(assert)
+	cdb, cleanup := prepareDatabase("read-document", assert)
+	defer cleanup()
 
 	// Create test document.
 	docA := MyDocument{
 		DocumentID: "foo-12345",
-		FieldA:     "foo",
-		FieldB:     12345,
+		Name:       "foo",
+		Age:        18,
 	}
 	resp := cdb.CreateDocument(docA)
 	assert.True(resp.IsOK())
@@ -198,8 +229,8 @@ func TestReadDocument(t *testing.T) {
 	err := resp.ResultValue(&docB)
 	assert.Nil(err)
 	assert.Equal(docB.DocumentID, docA.DocumentID)
-	assert.Equal(docB.FieldA, docA.FieldA)
-	assert.Equal(docB.FieldB, docA.FieldB)
+	assert.Equal(docB.Name, docA.Name)
+	assert.Equal(docB.Age, docA.Age)
 
 	// Try to read non-existant document.
 	resp = cdb.ReadDocument("i-do-not-exist")
@@ -210,13 +241,14 @@ func TestReadDocument(t *testing.T) {
 // TestUpdateDocument tests updating documents.
 func TestUpdateDocument(t *testing.T) {
 	assert := audit.NewTestingAssertion(t, true)
-	cdb := prepareDatabase(assert)
+	cdb, cleanup := prepareDatabase("update-document", assert)
+	defer cleanup()
 
 	// Create first revision.
 	docA := MyDocument{
 		DocumentID: "foo-12345",
-		FieldA:     "foo",
-		FieldB:     12345,
+		Name:       "foo",
+		Age:        22,
 	}
 	resp := cdb.CreateDocument(docA)
 	assert.True(resp.IsOK())
@@ -231,7 +263,7 @@ func TestUpdateDocument(t *testing.T) {
 	assert.Nil(err)
 
 	// Update the document.
-	docB.FieldB = 54321
+	docB.Age = 23
 
 	resp = cdb.UpdateDocument(docB)
 	assert.True(resp.IsOK())
@@ -244,8 +276,8 @@ func TestUpdateDocument(t *testing.T) {
 	assert.Nil(err)
 	assert.Equal(docC.DocumentID, docB.DocumentID)
 	assert.Substring("2-", docC.DocumentRevision)
-	assert.Equal(docC.FieldA, docB.FieldA)
-	assert.Equal(docC.FieldB, docB.FieldB)
+	assert.Equal(docC.Name, docB.Name)
+	assert.Equal(docC.Age, docB.Age)
 
 	// Read the first revision.
 	resp = cdb.ReadDocument(id, couchdb.Revision(revision))
@@ -254,19 +286,20 @@ func TestUpdateDocument(t *testing.T) {
 	err = resp.ResultValue(&docD)
 	assert.Nil(err)
 	assert.Equal(docD.DocumentRevision, revision)
-	assert.Equal(docD.FieldB, docA.FieldB)
+	assert.Equal(docD.Age, docA.Age)
 }
 
 // TestDeleteDocument tests deleting a document.
 func TestDeleteDocument(t *testing.T) {
 	assert := audit.NewTestingAssertion(t, true)
-	cdb := prepareDatabase(assert)
+	cdb, cleanup := prepareDatabase("delete-document", assert)
+	defer cleanup()
 
 	// Create test document.
 	docA := MyDocument{
 		DocumentID: "foo-12345",
-		FieldA:     "foo",
-		FieldB:     12345,
+		Name:       "foo",
+		Age:        33,
 	}
 	resp := cdb.CreateDocument(docA)
 	assert.True(resp.IsOK())
@@ -299,21 +332,58 @@ type MyDocument struct {
 	DocumentID       string `json:"_id,omitempty"`
 	DocumentRevision string `json:"_rev,omitempty"`
 
-	FieldA string
-	FieldB int
+	Name        string `json:"name"`
+	Age         int    `json:"age"`
+	Active      bool   `json:"active"`
+	Description string `json:"description"`
 }
 
-// prepareDatabase opens the database deletes a potention test
-// database and creates it newly.
-func prepareDatabase(assert audit.Assertion) couchdb.CouchDB {
-	cfg, err := etc.ReadString(TestingDBCfg)
+// prepareDatabase opens the database, deletes a possible test
+// database, and creates it newly.
+func prepareDatabase(database string, assert audit.Assertion) (couchdb.CouchDB, func()) {
+	cfgstr := strings.Replace(TemplateDBcfg, "<<DATABASE>>", database, 1)
+	cfg, err := etc.ReadString(cfgstr)
 	assert.Nil(err)
 	cdb, err := couchdb.Open(cfg)
 	assert.Nil(err)
 	resp := cdb.DeleteDatabase()
 	resp = cdb.CreateDatabase()
 	assert.True(resp.IsOK())
-	return cdb
+	return cdb, func() { cdb.DeleteDatabase() }
+}
+
+// prepareFilledDatabase opens the database, deletes a possible test
+// database, creates it newly and adds some data.
+func prepareFilledDatabase(database string, assert audit.Assertion) (couchdb.CouchDB, func()) {
+	cfgstr := strings.Replace(TemplateDBcfg, "<<DATABASE>>", database, 1)
+	cfg, err := etc.ReadString(cfgstr)
+	assert.Nil(err)
+	cdb, err := couchdb.Open(cfg)
+	assert.Nil(err)
+	resp := cdb.DeleteDatabase()
+	resp = cdb.CreateDatabase()
+	assert.True(resp.IsOK())
+
+	gen := audit.NewGenerator(audit.FixedRand())
+	docs := []interface{}{}
+	for i := 0; i < 1000; i++ {
+		first, middle, last := gen.Name()
+		doc := MyDocument{
+			DocumentID:  identifier.Identifier(last, first, i),
+			Name:        first + " " + middle + " " + last,
+			Age:         gen.Int(18, 65),
+			Active:      gen.FlipCoin(75),
+			Description: gen.Sentence(),
+		}
+		docs = append(docs, doc)
+	}
+	results, err := cdb.BulkWriteDocuments(docs...)
+	assert.Nil(err)
+	for _, result := range results {
+		assert.True(result.OK)
+	}
+
+	return cdb, func() { cdb.DeleteDatabase() }
 }
 
 // EOF