+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 
 ## 2016-11-21
 
+- Extended with bulk, design document, and view requests
+
+## 2016-11-21
+
 - Initial *alpha* with CRUD functionality